backend-fastapi/requirements.txt
fastapi
uvicorn
httpx
pydantic
google-generativeai==0.8.3
python-dotenv
motor
pymongo==4.6.1
beanie==1.26.0